Refactor presenter factories to use Kotlin Inject and update imports:
- Implement `Fake*PresenterFactory` classes for testing.
- Adjust affected tests and replace existing implementations with Fakes.
- Add `useKotlinInjectAnvilCompiler` to build configuration.
