+++ presenter/home/src/commonTest/kotlin/com/thomaskioko/tvmaniac/presenter/home/fakes/FakeWatchlistPresenterFactory.kt
@@ -1,0 +1,19 @@
+package com.thomaskioko.tvmaniac.presenter.home.fakes
+
+import com.arkivanov.decompose.ComponentContext
+import com.thomaskioko.tvmaniac.watchlist.presenter.WatchlistPresenter
+import com.thomaskioko.tvmaniac.watchlist.presenter.di.WatchlistPresenterFactory
+import com.thomaskioko.tvmaniac.watchlist.testing.FakeWatchlistRepository
+
+class FakeWatchlistPresenterFactory : WatchlistPresenterFactory {
+    val repository = FakeWatchlistRepository()
+
+    override fun create(
+        componentContext: ComponentContext,
+        navigateToShowDetails: (showDetails: Long) -> Unit,
+    ): WatchlistPresenter = WatchlistPresenter(
+        componentContext = componentContext,
+        navigateToShowDetails = navigateToShowDetails,
+        repository = repository,
+    )
+}
