+++ presenter/search/src/commonMain/kotlin/com/thomaskioko/tvmaniac/search/presenter/SearchShowsPresenter.kt
@@ -26,26 +26,6 @@ import me.tatarka.inject.annotations.Assisted
 import me.tatarka.inject.annotations.Inject
 
 @Inject
-class SearchPresenterFactory(
-    private val mapper: Mapper,
-    private val searchRepository: SearchRepository,
-    private val genreRepository: GenreRepository,
-) {
-    fun create(
-        componentContext: ComponentContext,
-        onNavigateToShowDetails: (id: Long) -> Unit,
-        onNavigateToGenre: (id: Long) -> Unit,
-    ): SearchShowsPresenter = SearchShowsPresenter(
-        componentContext = componentContext,
-        onNavigateToShowDetails = onNavigateToShowDetails,
-        onNavigateToGenre = onNavigateToGenre,
-        mapper = mapper,
-        searchRepository = searchRepository,
-        genreRepository = genreRepository,
-    )
-}
-
-@Inject
 class SearchShowsPresenter(
     @Assisted componentContext: ComponentContext,
     @Assisted private val onNavigateToShowDetails: (Long) -> Unit,
