+++ presenter/settings/build.gradle.kts
@@ -3,6 +3,10 @@ plugins {
 }
 
 tvmaniac {
+    multiplatform {
+        useKotlinInjectAnvilCompiler()
+    }
+
     optIn(
         "kotlinx.coroutines.ExperimentalCoroutinesApi",
     )
