+++ presenter/trailers/src/commonMain/kotlin/com/thomaskioko/tvmaniac/presenter/trailers/TrailersPresenter.kt
@@ -12,20 +12,6 @@ import kotlinx.coroutines.launch
 import me.tatarka.inject.annotations.Assisted
 import me.tatarka.inject.annotations.Inject
 
-@Inject
-class TrailersPresenterFactory(
-    private val repository: TrailerRepository,
-) {
-    fun create(
-        componentContext: ComponentContext,
-        id: Long,
-    ): TrailersPresenter = TrailersPresenter(
-        componentContext = componentContext,
-        traktShowId = id,
-        repository = repository,
-    )
-}
-
 class TrailersPresenter
 @Inject
 constructor(
