+++ presenter/watchlist/build.gradle.kts
@@ -3,6 +3,9 @@ plugins {
 }
 
 tvmaniac {
+    multiplatform {
+        useKotlinInjectAnvilCompiler()
+    }
     optIn(
         "kotlinx.coroutines.ExperimentalCoroutinesApi",
     )
